object-detector/requirements.txt
fastapi
uvicorn
torch
torchvision
Pillow
requests
python-multipart
kafka-python

